Refactored entityLayer packages to entity_layer
This change was pushed to main, but just to be safe I did it on this local branch, as well.
