+++ src/main/java/entity_layer/CommonSchedule.java
@@ -1,14 +1,13 @@
-package entityLayer;
+package entity_layer;
 
 /**
- * An interface for dealing with the CommonSchedule class, used as way to abstract our source files.
+ * A CommonSchedule class, implements the Schedule interface.
  * Created: 10/31/2022
  * Last updated: 11/4/2022
  *
  * @author MMachadoUofT
  */
-public interface Schedule {
-
+public class CommonSchedule implements Schedule {
     /**
      * Adds the passed TimeBlock to the availability.
      * If the given TimeBlock is contained within an existing availability entirely, nothing occurs.
@@ -17,10 +16,13 @@ public interface Schedule {
      *
      * @param availableBlock the TimeBlock to-be-added.
      */
-    void addAvailabilityBlock(TimeBlock availableBlock);
+    @Override
+    public void addAvailabilityBlock(TimeBlock availableBlock) {
+
+    }
 
     /**
-     * Removes the given TimeBlock from this Schedule's available TimeBlocks.
+     * Removes the given TimeBlock from this CommonSchedule's available TimeBlocks.
      * If the given TimeBlock does not overlap with any existing available TimeBlocks, nothing occurs.
      * If the given TimeBlock overlaps with existing available TimeBlocks, the TimeBlocks in question are shortened to
      * exclude the time held by this TimeBlock.
@@ -29,7 +31,10 @@ public interface Schedule {
      *
      * @param unavailableBlock the newly unavailable block of time.
      */
-    void removeAvailabilityBlock(TimeBlock unavailableBlock);
+    @Override
+    public void removeAvailabilityBlock(TimeBlock unavailableBlock) {
+
+    }
 
     /**
      * Returns the Curriculum holding the given Curriculum ID.
@@ -37,12 +42,18 @@ public interface Schedule {
      * @param curriculumID the ID belonging to the desired Curriculum object
      * @return the Curriculum object with the appropriate ID, null otherwise.
      */
-    Curriculum getCurriculum(int curriculumID);
+    @Override
+    public Curriculum getCurriculum(int curriculumID) {
+        return null;
+    }
 
     /**
      * Returns the TimeBlockManager object representing this Schedule's available times.
      *
      * @return this Schedule's available blocks, held by a TimeBlockManager object.
      */
-    TimeBlockManager getAvailability();
+    @Override
+    public TimeBlockManager getAvailability() {
+        return null;
+    }
 }
