+++ src/main/java/entity_layer/CommonTimeBlock.java
@@ -1,4 +1,4 @@
-package entityLayer;
+package entity_layer;
 
 import java.time.LocalDateTime;
