+++ src/main/java/entity_layer/Schedule.java
@@ -1,13 +1,14 @@
-package entityLayer;
+package entity_layer;
 
 /**
- * A CommonSchedule class, implements the Schedule interface.
+ * An interface for dealing with the CommonSchedule class, used as way to abstract our source files.
  * Created: 10/31/2022
  * Last updated: 11/4/2022
  *
  * @author MMachadoUofT
  */
-public class CommonSchedule implements Schedule {
+public interface Schedule {
+
     /**
      * Adds the passed TimeBlock to the availability.
      * If the given TimeBlock is contained within an existing availability entirely, nothing occurs.
@@ -16,13 +17,10 @@ public class CommonSchedule implements Schedule {
      *
      * @param availableBlock the TimeBlock to-be-added.
      */
-    @Override
-    public void addAvailabilityBlock(TimeBlock availableBlock) {
-
-    }
+    void addAvailabilityBlock(TimeBlock availableBlock);
 
     /**
-     * Removes the given TimeBlock from this CommonSchedule's available TimeBlocks.
+     * Removes the given TimeBlock from this Schedule's available TimeBlocks.
      * If the given TimeBlock does not overlap with any existing available TimeBlocks, nothing occurs.
      * If the given TimeBlock overlaps with existing available TimeBlocks, the TimeBlocks in question are shortened to
      * exclude the time held by this TimeBlock.
@@ -31,10 +29,7 @@ public class CommonSchedule implements Schedule {
      *
      * @param unavailableBlock the newly unavailable block of time.
      */
-    @Override
-    public void removeAvailabilityBlock(TimeBlock unavailableBlock) {
-
-    }
+    void removeAvailabilityBlock(TimeBlock unavailableBlock);
 
     /**
      * Returns the Curriculum holding the given Curriculum ID.
@@ -42,18 +37,12 @@ public class CommonSchedule implements Schedule {
      * @param curriculumID the ID belonging to the desired Curriculum object
      * @return the Curriculum object with the appropriate ID, null otherwise.
      */
-    @Override
-    public Curriculum getCurriculum(int curriculumID) {
-        return null;
-    }
+    Curriculum getCurriculum(int curriculumID);
 
     /**
      * Returns the TimeBlockManager object representing this Schedule's available times.
      *
      * @return this Schedule's available blocks, held by a TimeBlockManager object.
      */
-    @Override
-    public TimeBlockManager getAvailability() {
-        return null;
-    }
+    TimeBlockManager getAvailability();
 }
