+++ src/main/java/entity_layer/TimeBlock.java
@@ -1,4 +1,4 @@
-package entityLayer;
+package entity_layer;
 
 import java.time.LocalDateTime;
